bug in fixed huff table + bug in len table
The huff table fix was obvious, since the five bits are actually the
code for the distance table.

But the len table bug was evil - only the code 1 for the dist table was
wrong. Found it out while refactoring it to be a constexpr table lookup
instead of a function which uses a partial lookup.